[runtime-security] rename embedtests package name
