+++ pkg/security/embedtests/schemas.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/xeipuuv/gojsonschema"
 
-	"github.com/DataDog/datadog-agent/pkg/security/embed_tests/schemas"
+	"github.com/DataDog/datadog-agent/pkg/security/embedtests/schemas"
 	sprobe "github.com/DataDog/datadog-agent/pkg/security/probe"
 )
