+++ pkg/security/embedtests/syscalls_amd64.go
