+++ pkg/security/embedtests/trace_pipe.go
